backend-python/requirement.txt
uvicorn
fastapi
pydantic-settings
sqlalchemy
httpx
asyncpg
python-jose
passlib
pydantic[email]
google
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
aiofiles
boto3
celery
psycopg2-binary
python-multipart
pillow
pdf2image
google-cloud-vision
google-generativeai

